Moved out legacy code i will soon delete
