+++ server/src/main.rs
@@ -1,49 +1,32 @@
-use std::{net::SocketAddr, sync::Arc};
-
-
 use axum::{
     body::{boxed, Full},
     response::{IntoResponse, Response},
-    routing::{get, post},
+    routing::get,
     Router,
 };
 use axum_server::tls_rustls::RustlsConfig;
 use hyper::{header, StatusCode, Uri};
 use network_interface::{Addr, NetworkInterface, NetworkInterfaceConfig};
-use parking_lot::{const_mutex, Mutex};
 use rcgen::generate_simple_self_signed;
 use rust_embed::RustEmbed;
-use tokio::net::UdpSocket;
 use anyhow::Result;
 
 use crate::config::Config;
 
 mod config;
+mod legacy;
 
 #[tokio::main]
 async fn main() -> Result<()> {
     // Load config
     let config = Config::load()?;
 
-    // Setup udp
-    let sock = Arc::new(UdpSocket::bind("0.0.0.0:0".parse::<SocketAddr>()?).await?);
-    sock.connect(config.minion_udp_address).await?;
-
     // build our application with a single route
     let app = Router::new()
         .route("/", get(index_handler))
-        .route("/post_offer", post(post_offer))
-        .route("/get_offer", get(get_offer))
-        .route("/post_answer", post(post_answer))
-        .route("/get_answer", get(get_answer))
-        .route(
-            "/post_tracking",
-            post({
-                let sock = Arc::clone(&sock);
-                move |body| post_tracking(body, Arc::clone(&sock))
-            }),
-        )
         .fallback(get(static_handler));
+    
+    let app = app.merge(legacy::route(&config).await?);
 
     // Start the server based on config
     serve(app, &config).await?;
@@ -125,45 +108,4 @@ where
                 .unwrap(),
         }
     }
-}
-
-static OFFER: Mutex<String> = const_mutex(String::new());
-
-async fn post_offer(body: String) {
-    ANSWER.lock().clear();
-    let mut offer = OFFER.lock();
-    offer.clear();
-    offer.push_str(&body);
-}
-
-async fn get_offer() -> String {
-    let offer = OFFER.lock();
-    if offer.is_empty() {
-        "{}".to_string()
-    } else {
-        offer.clone()
-    }
-}
-
-static ANSWER: Mutex<String> = const_mutex(String::new());
-
-async fn post_answer(body: String) {
-    OFFER.lock().clear();
-    let mut answer = ANSWER.lock();
-    answer.clear();
-    answer.push_str(&body);
-}
-
-async fn get_answer() -> String {
-    let answer = ANSWER.lock();
-    if answer.is_empty() {
-        "{}".to_string()
-    } else {
-        answer.clone()
-    }
-}
-
-async fn post_tracking(body: String, sock: Arc<UdpSocket>) {
-    sock.send(body.as_bytes()).await.unwrap();
-    //println!("{}", body)
 }
